move tests into wizard_builder folder, thus avoiding import hell
