+++ .gitignore
@@ -2,7 +2,7 @@
 screendumps/
 
 # ignore test app db data
-tests/test_app/dump.json
+test-dump.json
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
