+++ wizard_builder/tests/urls.py
@@ -1,4 +1,4 @@
-from tests.test_app.views import edit_test_wizard_view, new_test_wizard_view
+from .test_app.views import edit_test_wizard_view, new_test_wizard_view
 
 from django.conf.urls import url
 from django.contrib import admin
